Add collider component

+++ KEngine/Inc/collision_manager.h
@@ -10,10 +10,10 @@ namespace K
 	{
 		Vector3 min;
 		Vector3 max;
-		uint8_t x_count;
-		uint8_t y_count;
-		uint8_t z_count;
-		uint32_t total_count;
+		int x_count;
+		int y_count;
+		int z_count;
+		int total_count;
 		Vector3 section_diagonal_length;
 		std::unique_ptr<CollisionSection[], std::function<void(CollisionSection*)>> section;
 	};
@@ -41,7 +41,7 @@ namespace K
 
 		virtual void _Finalize() override;
 
-		void _CreateCollisionGroup(std::string const& _tag, Vector3 const& _min, Vector3 const& _max, uint8_t _x_count, uint8_t _y_count, uint8_t _z_count);
+		void _CreateCollisionGroup(std::string const& _tag, Vector3 const& _min, Vector3 const& _max, int _x_count, int _y_count, int _z_count);
 
 		std::unordered_map<std::string, std::unique_ptr<CollisionGroup>> collision_group_map_{};
 	};
